fix: fix typo

+++ backend/pkg/repository/database/integration/dao_integration_config.go
@@ -229,7 +229,7 @@ WHERE rn = 1`
 
 	latestAPI := make(map[string]any)
 	var apmList []string
-	var maxTimeout int64 = 0
+	var maxTimeout int = 0
 	for _, api := range latestTraceAPIs {
 		var apiSpec map[string]interface{}
 		// 反序列化 JSON 数据到 map
@@ -246,8 +246,8 @@ WHERE rn = 1`
 		latestAPI[api.ApmType] = cfg
 		apmList = append(apmList, api.ApmType)
 		timeout, ok := apiSpec["timeout"]
-		if ok && timeout.(int64) > maxTimeout {
-			maxTimeout = timeout.(int64)
+		if ok && timeout.(int) > maxTimeout {
+			maxTimeout = timeout.(int)
 		}
 	}
